Rework the GET/PUT for clients.
The client code now responds to reconfiguration changes, and server
failures, and appropriately notifies the client.  Before, it would just
sit there without reacting to either reconfiguration or failure.

+++ src/hyperdex/replication_manager.cc
@@ -1340,7 +1340,7 @@ hyperdex :: replication_manager :: respond_to_client(clientop co,
                                                      network_returncode ret)
 {
     e::buffer msg(5);
-    uint8_t result = static_cast<uint8_t>(ret);
+    uint16_t result = static_cast<uint16_t>(ret);
     msg.pack() << co.nonce << result;
     m_clientops.remove(co);
     m_comm->send(co.region, co.from, type, msg);
